pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "amavrits"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "torch (>=2.7.1,<3.0.0)",
    "xgboost (>=3.0.2,<4.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "yfinance (>=0.2.63,<0.3.0)",
    "vectorbt (>=0.27.3,<0.28.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "ta (>=0.11.0,<0.12.0)",
    "typer (>=0.16.0,<0.17.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
